regulate types in configs

+++ src/lib/client_conn_group.rs
@@ -23,7 +23,7 @@ pub enum ConnectionGroupError {
 #[derive(Debug, Clone, Copy)]
 pub struct Config {
     pub max_packet_ahead: u16,
-    pub max_packet_size: usize,
+    pub max_packet_size: u16,
 }
 
 impl Into<session::server::Config> for Config {
