+++ src/lib/session/client.rs
@@ -7,7 +7,7 @@ use crate::prelude::*;
 #[derive(Debug, Clone, Copy)]
 pub struct Config {
     pub max_packet_ahead: u16,
-    pub max_packet_size: usize,
+    pub max_packet_size: u16,
 }
 
 impl Into<session::sequenced_to_stream::Config> for Config {
